Fix: Ids del Requestor pueden ser String
Hasta ahora todos los campos del Requestor de DataService eran de tipo
Long, pero se enviaban siempre a null y se recibían a null. Con cambios
recientes de servidor aunque se mande a null viene relleno, así que se
cambian los tipos del Requestor para admitir cualquier dato incluyendo
String, o fallará el mapeo de datos.

+++ app/src/main/java/com/shootr/android/service/dataservice/ShootrDataService.java
@@ -644,7 +644,7 @@ public class ShootrDataService implements ShootrService {
     }
 
     private void updateTimeFromServer(GenericDto dto) {
-        Long serverTime = dto.getRequestor().getReq()[RequestorDto.POSITION_SYSTEM_TIME];
+        Long serverTime = (Long) dto.getRequestor().getReq()[RequestorDto.POSITION_SYSTEM_TIME];
         if (serverTime != null) {
             timeUtils.setCurrentTime(serverTime);
         }
